fixed const

+++ src/fable/Fable.Core/AST/AST.Fable.Util.fs
@@ -64,8 +64,8 @@ let makeConst (value: obj) =
     | :? uint32 as x -> NumberConst (U2.Case1 (int x), UInt32)
     // Float types
     | :? float as x -> NumberConst (U2.Case2 x, Float64)
-    | :? int64 as x -> NumberConst (U2.Case2 (float x), Int64)
-    | :? uint64 as x -> NumberConst (U2.Case2 (float x), UInt64)
+    | :? int64 as x -> NumberConst (U2.Case2 (float x), Float64)
+    | :? uint64 as x -> NumberConst (U2.Case2 (float x), Float64)
     | :? float32 as x -> NumberConst (U2.Case2 (float x), Float32)
     | :? decimal as x -> NumberConst (U2.Case2 (float x), Float64)
     // TODO: Regex
